Fixed DebugConsole not deleting itself
The debug console was the problem with issue #44. I fixed it by inheriting the Singleton class.  I had to change how
the application generated itself however, which led to the creation of the CreateInstance().

When DisplayConsole is called now, it will check if the instance is null, and if it is, it'll make sure
CreateInstance() is called so all the parts can be generated properly, since the object otherwise doesn't exist at
Runtime.