fix #52 replaced Refugee Health with Shelter
Replaced the code in the refugee script to utilize Shelter instead. I
took into account Tejas' code for the lose/win state, which rely on the
OnDiedHandler function, so, instead of killing this function, I now just
use this as a function call so it does not break his stuff.