WIP1 Archteure inital
Created an inital setup for the architecture for WIP1.  There needs to be more agregation lines in place, but every package has an overview model and a type heirarchy model which helps breaks things down.

There's still a need to see if we can cross reference parts between the different classes easily or not.  Don't want ot manually reference everything by hand....